+++ R/np.plot.engine.plbandwidth.R
@@ -225,7 +225,8 @@
           bws = bws,
           txdat = xdat,
           tydat = ydat,
-          tzdat = zdat
+          tzdat = zdat,
+          se = TRUE
         )
       } else {
         .np_plot_plreg_local_fit(
@@ -233,7 +234,8 @@
           xdat = xdat,
           ydat = ydat,
           zdat = zdat,
-          .np.empty.report = .np.empty.report
+          .np.empty.report = .np.empty.report,
+          se = plot.errors || plot.behavior != "plot"
         )
       }
       cf <- as.double(fit.coef$xcoef)
@@ -355,7 +357,8 @@
           zdat = zdat,
           exdat = x.eval[,1, drop = FALSE],
           ezdat = x.eval[,2, drop = FALSE],
-          .np.empty.report = .np.empty.report
+          .np.empty.report = .np.empty.report,
+          se = plot.behavior != "plot"
         )
       }
 
@@ -745,7 +748,8 @@
             zdat = zdat,
             exdat = subcol(exdat,ei,i)[seq_len(xi.neval),, drop = FALSE],
             ezdat = ezdat[seq_len(xi.neval),, drop = FALSE],
-            .np.empty.report = .np.empty.report
+            .np.empty.report = .np.empty.report,
+            se = plot.behavior != "plot"
           )
         }
 
@@ -1052,7 +1056,8 @@
             zdat = zdat,
             exdat = exdat[seq_len(xi.neval),, drop = FALSE],
             ezdat = subcol(ezdat,ei,i)[seq_len(xi.neval),, drop = FALSE],
-            .np.empty.report = .np.empty.report
+            .np.empty.report = .np.empty.report,
+            se = plot.behavior != "plot"
           )
         }